Refactor out ModelStub
