+++ src/test/java/seedu/address/logic/commands/AddPairCommandTest.java
@@ -47,7 +47,7 @@ public class AddPairCommandTest {
     public void execute_duplicatePair_throwsCommandException() {
         Pair validPair = new PairBuilder().build();
         AddPairCommand addPairCommand = new AddPairCommand(validPair);
-        ModelStub modelStub = new ModelStubWithPair(validPair);
+        CommandTestUtil.ModelStub modelStub = new ModelStubWithPair(validPair);
 
         assertThrows(CommandException.class,
                 AddPairCommand.MESSAGE_DUPLICATE_PAIR, () -> addPairCommand.execute(modelStub));
@@ -80,114 +80,9 @@ public class AddPairCommandTest {
     }
 
     /**
-     * A default model stub that have all of the methods failing.
-     */
-    private class ModelStub implements Model {
-        @Override
-        public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyUserPrefs getUserPrefs() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public GuiSettings getGuiSettings() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setGuiSettings(GuiSettings guiSettings) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getFriendlyLinkFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setFriendlyLinkFilePath(Path addressBookFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addPerson(Person person) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setFriendlyLink(ReadOnlyFriendlyLink newData) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyFriendlyLink getFriendlyLink() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasPerson(Person person) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deletePerson(Person target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setPerson(Person target, Person editedPerson) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addPair(Pair pair) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasPair(Pair pair) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deletePair(Pair target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setPair(Pair target, Pair editedPair) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Person> getFilteredPersonList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredPersonList(Predicate<Person> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Pair> getFilteredPairList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredPairList(Predicate<Pair> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-    }
-
-    /**
      * A Model stub that contains a single pair.
      */
-    private class ModelStubWithPair extends ModelStub {
+    private class ModelStubWithPair extends CommandTestUtil.ModelStub {
         private final Pair pair;
 
         ModelStubWithPair(Pair pair) {
@@ -205,7 +100,7 @@ public class AddPairCommandTest {
     /**
      * A Model stub that always accept the pair being added.
      */
-    private class ModelStubAcceptingPairAdded extends ModelStub {
+    private class ModelStubAcceptingPairAdded extends CommandTestUtil.ModelStub {
         final ArrayList<Pair> pairsAdded = new ArrayList<>();
 
         @Override
